feature: home page made dynamic

+++ admin/delete-categories.php
@@ -6,10 +6,19 @@ if(isset($_GET['id'])) {
     $id = filter_var($_GET['id'], FILTER_SANITIZE_NUMBER_INT);
 
 
-    // update  categories id posts that belongs to this category to id of uncategorized category
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Category deleted successfully";
+    // for later
+    // update category_id of posts that belongs to this category to id of uncaegorized category
+    $update_query = "UPDATE posts SET category_id=6 WHERE category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
+
+    if(!mysqli_errno($connection)) {
+        // delete category from database
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Category deleted successfully";
+    }
+
+
 }
 
 header('location:manage-categories.php');
